pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "microscape"
version = "0.1.0"
description = "Downstream analysis tools for amplicon sequencing data, designed to work with papa2's output."
readme = "README.md"
license = "BSD-3-Clause"
requires-python = ">=3.9"
authors = [
    { name = "rec3141" },
]
keywords = ["amplicon", "metagenomics", "16S", "18S", "microbiome"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "numpy",
    "pandas",
    "scipy",
    "scikit-learn",
    "matplotlib",
]

[project.optional-dependencies]
bio = [
    "papa2",
    "biopython",
]

[tool.setuptools.packages.find]
include = ["microscape*"]
